pyproject.toml
[project]
name = "flask-web-service"
version = "0.1.0"
description = "A Flask web service for database interaction"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.8.1"
dependencies = [
    "cryptography>=45.0.3",
    "flask>=2.3.0",
    "flask-sqlalchemy>=3.0.0",
    "flask-jwt-extended>=4.5.0",
    "flask-cors>=4.0.0",
    "flask-migrate>=4.0.0",
    "flask-limiter>=3.5.0",
    "pymysql>=1.1.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "sshtunnel>=0.4.0",
    "ulid-py>=1.1.0",
    "werkzeug>=2.3.0",
    "bcrypt>=4.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "black>=22.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
